refactor(project): rework files hierarchy + format components
