+++ src/assets/images/atomic-design.png
