+++ src/index.ts
@@ -1,1 +1,1 @@
-export * from './ui/index'
+export * from '@ui/index'
