+++ src/ui/styles/themes.scss
@@ -1,4 +1,4 @@
-@import '../styles/palette.scss';
+@import './palette.scss';
 
 $themes: (
   light: (
